feat: add more components to accessibility checks view (#1388)
Add more content to the accessibility checks view. 

Not part of this PR:
- hook up the Cards UI
- handle the "scanning" state
