+++ src/electron/views/automated-checks/components/command-bar.scss
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 @import '../../../../common/styles/colors.scss';
 
-.automated-checks-command-bar {
+.command-bar {
     width: 100%;
     display: flex;
     justify-content: space-between;
